requirements.txt
adagio==0.2.4
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astropy==5.3
astroquery==0.4.7.dev8738
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autoviz==0.1.604
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==2.4.3
boto3==1.26.142
botocore==1.29.142
catboost==1.2
category-encoders==2.6.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.6
colorcet==3.0.1
colorlog==6.7.0
colour==0.1.5
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.35
daal==2023.1.1
daal4py==2023.1.1
dacite==1.8.1
dash==2.10.0
dash-auth==2.0.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dask==2023.5.1
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
deepchecks==0.14.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
distributed==2023.5.1
docker==6.1.2
dtreeviz==2.2.1
emoji==2.4.0
entrypoints==0.4
evidently==0.2.8
exceptiongroup==1.1.1
explainerdashboard==0.4.2.2
fairlearn==0.7.0
fastapi==0.95.2
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.0
Flask==2.2.3
flask-simplelogin==0.1.2
Flask-WTF==1.1.1
fonttools==4.39.4
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.5.0
fugue==0.8.4
fugue-sql-antlr==0.1.6
future==0.18.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
gradio==3.32.0
gradio_client==0.2.5
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
holoviews==1.14.9
html5lib==1.1
htmlmin==0.1.12
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
hvplot==0.8.3
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.30.0
imbalanced-learn==0.10.1
importlib-metadata==5.2.0
importlib-resources==5.12.0
iniconfig==2.0.0
interpret==0.4.1
interpret-core==0.4.1
ipykernel==6.23.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-server==1.24.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kaleido==0.2.1
keyring==24.2.0
kiwisolver==1.4.4
kmodes==0.12.2
lazy_loader==0.2
lightgbm==3.3.5
lime==0.2.0.1
linkify-it-py==2.0.2
llvmlite==0.40.0
locket==1.0.0
lxml==4.9.2
m2cgen==0.10.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mlflow==1.30.1
mlxtend==0.22.0
more-itertools==10.1.0
moto==4.1.10
msgpack==1.0.5
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.23.0
oauthlib==3.2.2
optuna==3.2.0
orjson==3.8.14
outcome==1.2.0
oyaml==1.0
packaging==21.3
pandas==1.5.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.4.0
patsy==0.5.3
percy==2.0.2
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
plotly-resampler==0.8.3.2
pluggy==1.0.0
pmdarima==2.0.3
prometheus-client==0.17.0
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
py4j==0.10.9.7
pyamg==5.0.0
pyaml==23.5.9
pyarrow==12.0.0
pycaret==3.0.2
pycparser==2.21
pyct==0.5.0
pydantic==1.10.8
pydub==0.25.1
pyerfa==2.0.0.3
Pygments==2.15.1
PyJWT==2.7.0
pynndescent==0.5.10
PyNomaly==0.3.3
pyod==1.0.9
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-utils==3.6.0
pytz==2022.7.1
pyviz-comms==2.3.0
pyvo==1.4.2
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==23.2.1
qpd==0.4.1
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
ray==2.4.0
regex==2023.5.5
requests==2.31.0
responses==0.23.1
retrying==1.3.4
s3transfer==0.6.1
SALib==1.4.7
schemdraw==0.15
scikit-image==0.20.0
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.9.1
seaborn==0.12.2
selenium==4.2.0
semantic-version==2.10.0
Send2Trash==1.8.2
shap==0.42.1
six==1.16.0
skope-rules==1.0.1
sktime==0.17.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlglot==14.1.1
sqlparse==0.4.4
starlette==0.27.0
statsforecast==1.5.0
statsmodels==0.14.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tbb==2021.9.0
tblib==1.7.0
tenacity==8.2.2
tensorboardX==2.6
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.9.0
treeinterpreter==0.2.3
triad==0.8.8
trio==0.22.0
trio-websocket==0.10.2
tune-sklearn==0.4.5
typeguard==2.13.3
types-PyYAML==6.0.12.10
typing_extensions==4.6.2
uc-micro-py==1.0.2
umap-learn==0.5.3
urllib3==1.26.16
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
virtualenv==20.21.0
visions==0.7.5
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==3.6.4
wordcloud==1.9.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.5
xlrd==2.0.1
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
ydata-profiling==4.2.0
yellowbrick==1.5
zict==3.0.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
